NewFeature: Split out DiffractionVector1D and DiffractionVector2D to better align with hyperspy Signal definition.
